+++ config/configload/load.go
@@ -29,12 +29,14 @@ const (
 	definitions  = "definitions"
 	endpoint     = "endpoint"
 	errorHandler = "error_handler"
+	files        = "files"
 	nameLabel    = "name"
 	oauth2       = "oauth2"
 	proxy        = "proxy"
 	request      = "request"
 	server       = "server"
 	settings     = "settings"
+	spa          = "spa"
 	// defaultNameLabel maps the hcl label attr 'name'.
 	defaultNameLabel = "default"
 )
@@ -384,6 +386,32 @@ func LoadConfig(body hcl.Body, src []byte, filename, dirPath string) (*config.Co
 			if err = configureErrorHandler(apiErrorHandler, helper); err != nil {
 				return nil, err
 			}
+		}
+
+		for _, spaBlock := range bodyToContent(serverConfig.Remain).Blocks.OfType(spa) {
+			spaConfig := &config.Spa{}
+			if diags := gohcl.DecodeBody(spaBlock.Body, helper.context, spaConfig); diags.HasErrors() {
+				return nil, diags
+			}
+
+			if len(spaBlock.Labels) > 0 {
+				spaConfig.Name = spaBlock.Labels[0]
+			}
+
+			serverConfig.SPAs = append(serverConfig.SPAs, spaConfig)
+		}
+
+		for _, filesBlock := range bodyToContent(serverConfig.Remain).Blocks.OfType(files) {
+			filesConfig := &config.Files{}
+			if diags := gohcl.DecodeBody(filesBlock.Body, helper.context, filesConfig); diags.HasErrors() {
+				return nil, diags
+			}
+
+			if len(filesBlock.Labels) > 0 {
+				filesConfig.Name = filesBlock.Labels[0]
+			}
+
+			serverConfig.Files = append(serverConfig.Files, filesConfig)
 		}
 
 		// standalone endpoints