+++ config/configload/merge.go
@@ -22,20 +22,51 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 	type (
 		namedBlocks   map[string]*hclsyntax.Block
 		apiDefinition struct {
-			labels       []string
-			attributes   hclsyntax.Attributes
-			blocks       namedBlocks
-			endpoints    namedBlocks
-			errorHandler namedBlocks
+			labels          []string
+			typeRange       hcl.Range
+			labelRanges     []hcl.Range
+			openBraceRange  hcl.Range
+			closeBraceRange hcl.Range
+			attributes      hclsyntax.Attributes
+			blocks          namedBlocks
+			endpoints       namedBlocks
+			errorHandler    namedBlocks
+		}
+		spaDefinition struct {
+			labels          []string
+			typeRange       hcl.Range
+			labelRanges     []hcl.Range
+			openBraceRange  hcl.Range
+			closeBraceRange hcl.Range
+			attributes      hclsyntax.Attributes
+			blocks          namedBlocks
+		}
+		filesDefinition struct {
+			labels          []string
+			typeRange       hcl.Range
+			labelRanges     []hcl.Range
+			openBraceRange  hcl.Range
+			closeBraceRange hcl.Range
+			attributes      hclsyntax.Attributes
+			blocks          namedBlocks
 		}
 		namedAPIs        map[string]*apiDefinition
+		namedSPAs        map[string]*spaDefinition
+		namedFiles       map[string]*filesDefinition
 		serverDefinition struct {
-			labels     []string
-			attributes hclsyntax.Attributes
-			blocks     namedBlocks
-			endpoints  namedBlocks
-			apis       namedAPIs
+			labels          []string
+			typeRange       hcl.Range
+			labelRanges     []hcl.Range
+			openBraceRange  hcl.Range
+			closeBraceRange hcl.Range
+			attributes      hclsyntax.Attributes
+			blocks          namedBlocks
+			endpoints       namedBlocks
+			apis            namedAPIs
+			spas            namedSPAs
+			files           namedFiles
 		}
+
 		servers map[string]*serverDefinition
 	)
 
@@ -49,6 +80,14 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 				blocks[<name>]       = hclsyntax.Block (cors)
 				endpoints[<key>]     = hclsyntax.Block
 				error_handler[<key>] = hclsyntax.Block
+			}
+			spas[<key>]      = {
+				attributes           = hclsyntax.Attributes
+				blocks[<name>]       = hclsyntax.Block (cors)
+			}
+			files[<key>]      = {
+				attributes           = hclsyntax.Attributes
+				blocks[<name>]       = hclsyntax.Block (cors)
 			}
 		}
 	*/
@@ -82,11 +121,17 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 
 			if results[serverKey] == nil {
 				results[serverKey] = &serverDefinition{
-					labels:     outerBlock.Labels,
-					attributes: make(hclsyntax.Attributes),
-					blocks:     make(namedBlocks),
-					endpoints:  make(namedBlocks),
-					apis:       make(namedAPIs),
+					labels:          outerBlock.Labels,
+					typeRange:       outerBlock.TypeRange,
+					labelRanges:     outerBlock.LabelRanges,
+					openBraceRange:  outerBlock.OpenBraceRange,
+					closeBraceRange: outerBlock.CloseBraceRange,
+					attributes:      make(hclsyntax.Attributes),
+					blocks:          make(namedBlocks),
+					endpoints:       make(namedBlocks),
+					apis:            make(namedAPIs),
+					spas:            make(namedSPAs),
+					files:           make(namedFiles),
 				}
 			}
 
@@ -99,14 +144,11 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 			}
 
 			for _, block := range outerBlock.Body.Blocks {
-				uniqueAPILabels := make(map[string]struct{})
+				uniqueAPILabels, uniqueSPALabels, uniqueFilesLabels := make(map[string]struct{}), make(map[string]struct{}), make(map[string]struct{})
 
-				// TODO: Do we need this IF around the FOR?
-				if block.Type == "files" || block.Type == "spa" || block.Type == api || block.Type == endpoint {
-					for _, name := range []string{"error_file", "document_root"} {
-						if attr, ok := block.Body.Attributes[name]; ok {
-							block.Body.Attributes[name].Expr = absPath(attr)
-						}
+				for _, name := range []string{"error_file", "document_root"} {
+					if attr, ok := block.Body.Attributes[name]; ok {
+						block.Body.Attributes[name].Expr = absPath(attr)
 					}
 				}
 
@@ -156,11 +198,15 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 
 					if results[serverKey].apis[apiKey] == nil {
 						results[serverKey].apis[apiKey] = &apiDefinition{
-							labels:       block.Labels,
-							attributes:   make(hclsyntax.Attributes),
-							blocks:       make(namedBlocks),
-							endpoints:    make(namedBlocks),
-							errorHandler: make(namedBlocks),
+							labels:          block.Labels,
+							typeRange:       block.TypeRange,
+							labelRanges:     block.LabelRanges,
+							openBraceRange:  block.OpenBraceRange,
+							closeBraceRange: block.CloseBraceRange,
+							attributes:      make(hclsyntax.Attributes),
+							blocks:          make(namedBlocks),
+							endpoints:       make(namedBlocks),
+							errorHandler:    make(namedBlocks),
 						}
 					}
 
@@ -190,6 +236,80 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 						} else {
 							results[serverKey].apis[apiKey].blocks[subBlock.Type] = subBlock
 						}
+					}
+				} else if block.Type == spa {
+					var spaKey string
+
+					if len(block.Labels) > 0 {
+						spaKey = block.Labels[0]
+					}
+
+					if len(bodies) > 1 {
+						if _, ok := uniqueSPALabels[spaKey]; ok {
+							return nil, newMergeError(errUniqueLabels, block)
+						}
+
+						uniqueSPALabels[spaKey] = struct{}{}
+					} else {
+						// Create unique key for multiple spa blocks inside a single config file.
+						spaKey += fmt.Sprintf("|%p", &spaKey)
+					}
+
+					if results[serverKey].spas[spaKey] == nil {
+						results[serverKey].spas[spaKey] = &spaDefinition{
+							labels:          block.Labels,
+							typeRange:       block.TypeRange,
+							labelRanges:     block.LabelRanges,
+							openBraceRange:  block.OpenBraceRange,
+							closeBraceRange: block.CloseBraceRange,
+							attributes:      make(hclsyntax.Attributes),
+							blocks:          make(namedBlocks),
+						}
+					}
+
+					for name, attr := range block.Body.Attributes {
+						results[serverKey].spas[spaKey].attributes[name] = attr
+					}
+
+					for _, subBlock := range block.Body.Blocks {
+						results[serverKey].spas[spaKey].blocks[subBlock.Type] = subBlock
+					}
+				} else if block.Type == files {
+					var filesKey string
+
+					if len(block.Labels) > 0 {
+						filesKey = block.Labels[0]
+					}
+
+					if len(bodies) > 1 {
+						if _, ok := uniqueFilesLabels[filesKey]; ok {
+							return nil, newMergeError(errUniqueLabels, block)
+						}
+
+						uniqueFilesLabels[filesKey] = struct{}{}
+					} else {
+						// Create unique key for multiple files blocks inside a single config file.
+						filesKey += fmt.Sprintf("|%p", &filesKey)
+					}
+
+					if results[serverKey].files[filesKey] == nil {
+						results[serverKey].files[filesKey] = &filesDefinition{
+							labels:          block.Labels,
+							typeRange:       block.TypeRange,
+							labelRanges:     block.LabelRanges,
+							openBraceRange:  block.OpenBraceRange,
+							closeBraceRange: block.CloseBraceRange,
+							attributes:      make(hclsyntax.Attributes),
+							blocks:          make(namedBlocks),
+						}
+					}
+
+					for name, attr := range block.Body.Attributes {
+						results[serverKey].files[filesKey].attributes[name] = attr
+					}
+
+					for _, subBlock := range block.Body.Blocks {
+						results[serverKey].files[filesKey].blocks[subBlock.Type] = subBlock
 					}
 				} else {
 					results[serverKey].blocks[block.Type] = block
@@ -233,9 +353,59 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 					Attributes: apiBlock.attributes,
 					Blocks:     apiBlocks,
 				},
+				TypeRange:       apiBlock.typeRange,
+				LabelRanges:     apiBlock.labelRanges,
+				OpenBraceRange:  apiBlock.openBraceRange,
+				CloseBraceRange: apiBlock.closeBraceRange,
 			}
 
 			serverBlocks = append(serverBlocks, mergedAPI)
+		}
+
+		for _, spaBlock := range serverBlock.spas {
+			var spaBlocks hclsyntax.Blocks
+
+			for _, b := range spaBlock.blocks {
+				spaBlocks = append(spaBlocks, b)
+			}
+
+			mergedSPA := &hclsyntax.Block{
+				Type:   spa,
+				Labels: spaBlock.labels,
+				Body: &hclsyntax.Body{
+					Attributes: spaBlock.attributes,
+					Blocks:     spaBlocks,
+				},
+				TypeRange:       spaBlock.typeRange,
+				LabelRanges:     spaBlock.labelRanges,
+				OpenBraceRange:  spaBlock.openBraceRange,
+				CloseBraceRange: spaBlock.closeBraceRange,
+			}
+
+			serverBlocks = append(serverBlocks, mergedSPA)
+		}
+
+		for _, filesBlock := range serverBlock.files {
+			var filesBlocks hclsyntax.Blocks
+
+			for _, b := range filesBlock.blocks {
+				filesBlocks = append(filesBlocks, b)
+			}
+
+			mergedFiles := &hclsyntax.Block{
+				Type:   files,
+				Labels: filesBlock.labels,
+				Body: &hclsyntax.Body{
+					Attributes: filesBlock.attributes,
+					Blocks:     filesBlocks,
+				},
+				TypeRange:       filesBlock.typeRange,
+				LabelRanges:     filesBlock.labelRanges,
+				OpenBraceRange:  filesBlock.openBraceRange,
+				CloseBraceRange: filesBlock.closeBraceRange,
+			}
+
+			serverBlocks = append(serverBlocks, mergedFiles)
 		}
 
 		mergedServer := &hclsyntax.Block{
@@ -245,6 +415,10 @@ func mergeServers(bodies []*hclsyntax.Body) (hclsyntax.Blocks, error) {
 				Attributes: serverBlock.attributes,
 				Blocks:     serverBlocks,
 			},
+			TypeRange:       serverBlock.typeRange,
+			LabelRanges:     serverBlock.labelRanges,
+			OpenBraceRange:  serverBlock.openBraceRange,
+			CloseBraceRange: serverBlock.closeBraceRange,
 		}
 
 		mergedServers = append(mergedServers, mergedServer)