+++ server/http_test.go
@@ -77,7 +77,7 @@ func TestHTTPServer_ServeHTTP_Files(t *testing.T) {
 	srvConf, err := runtime.NewServerConfiguration(conf, logger, tmpMemStore)
 	helper.Must(err)
 
-	spaContent, err := os.ReadFile(conf.Servers[0].Spa.BootstrapFile)
+	spaContent, err := os.ReadFile(conf.Servers[0].SPAs[0].BootstrapFile)
 	helper.Must(err)
 
 	port := runtime.Port(conf.Settings.DefaultPort)
@@ -167,7 +167,7 @@ func TestHTTPServer_ServeHTTP_Files2(t *testing.T) {
 	helper.Must(err)
 
 	error404Content := []byte("<html><body><h1>route not found error: My custom error template</h1></body></html>")
-	spaContent, err := os.ReadFile(conf.Servers[0].Spa.BootstrapFile)
+	spaContent, err := os.ReadFile(conf.Servers[0].SPAs[0].BootstrapFile)
 	helper.Must(err)
 
 	tmpStoreCh := make(chan struct{})
@@ -643,7 +643,7 @@ func TestHTTPServer_RequestID(t *testing.T) {
 			}
 
 			if tc.status == http.StatusOK {
-				if lastLog.Message != "" {
+				if lastLog != nil && lastLog.Message != "" {
 					subT.Errorf("Unexpected log message given: %s", lastLog.Message)
 				}